requirements.txt
fastapi
uvicorn
websockets
kafka-python
redis
starlette
itsdangerous
pymongo
bcrypt